# contracts/openapi/bff-v1.yaml
openapi: 3.0.3
info:
  title: UnoArena Realtime Gateway / BFF
  version: 1.0.0
  description: |
    Sole external HTTP/SSE surface for UnoArena. Clients submit versioned command
    envelopes, fetch reconnect/projection snapshots, and subscribe to SSE streams.
    Microservices are not publicly reachable. For this implementation the repo-owned
    simple CLI is the sole client/test interface; a graphical UI is deferred and must
    continue to use only this BFF boundary. Rejected commands return `status=rejected`
    with a structured operational/security audit record only (never domain events or
    Game Integrity entries). Unknown or evicted `Last-Event-ID` values return HTTP 409
    `snapshot_required` so clients resync via the snapshot routes before resuming SSE.
servers:
  - url: http://localhost:8080
    description: Local compose BFF (gateway)
...
paths:
...
  /v1/rankings/leaderboards:
    get:
      tags: [reads]
      summary: Public leaderboard read model (proxied)
      operationId: getLeaderboard
      parameters:
        - $ref: "#/components/parameters/CorrelationIdHeader"
        - name: boardType
          in: query
          required: true
          schema:
            type: string
            enum: [casual_elo, tournament_placement]
        - name: cursor
          in: query
          required: false
          schema:
            type: string
          description: Opaque Ranking-owned live keyset cursor over the last (rating, playerId) boundary. A projection version change does not invalidate it.
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 100
            minimum: 1
            maximum: 500
      responses:
        "200":
          description: Leaderboard snapshot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeaderboardPage"
        "429":
          $ref: "#/components/responses/Error"
        "502":
          description: Ranking upstream failure
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
...
components:
  securitySchemes:
    SessionBearer:
      type: http
      scheme: bearer
      bearerFormat: opaque-session-token
      description: |
        Identity-issued session token via BFF login. Required for player/command/control
        routes. Optional on spectator routes; when present, invalid tokens yield 401.
  parameters:
    CorrelationIdHeader:
      name: X-Correlation-Id
      in: header
      required: false
      schema:
        type: string
...
  responses:
    Error:
      description: Structured error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"
...
  schemas:
...
    LeaderboardPage:
      type: object
      additionalProperties: false
      required: [boardType, projectionVersion, generatedAt, entries]
      properties:
        boardType:
          type: string
          enum: [casual_elo, tournament_placement]
        projectionVersion:
          type: integer
          format: int64
          minimum: 0
          description: Version observed for this page; a later page may report a newer version.
        generatedAt:
          type: string
          format: date-time
        entries:
          type: array
          maxItems: 500
          items:
            type: object
            additionalProperties: false
            required: [playerId, rating, rank]
            properties:
              playerId:
                type: string
              rating:
                type: integer
              rank:
                type: integer
                minimum: 1
        nextCursor:
          type: string
          description: Opaque live-keyset continuation cursor; absent on the final page.
...
    ErrorBody:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        correlationId:
          type: string
        commandId:
          type: string
...